FRONTEND-28: show files in app details page
